requirements.txt
torch
torchao
timm
numpy
Pillow
matplotlib
tqdm
editdistance
optuna
onnxscript
onnxruntime
opencv-python
albumentations